Feature: make-pretty for all languages / Updates: Github yml build files and make-pretty applied (#169)
This updates the `make-pretty` script to support all used languages in the project. It's based on the same script in the `openthread` repo. Also make-pretty is applied on all sources.

Additionally, Github yml files are updated to better comply with recommended steps order and to bump setup-go to v5. Some steps are renamed to provide more clarity what it does. And the 'build' test is now split in the Go build part and the OT nodes build part. These 2 parts each require a separate matrix (since the nodes build doesn't involve Go, and the Go build doesn't involve C/C++ )

* [ot-rfsim] make-pretty clang-tidy
* [ot-rfsim] make-pretty clang-format
* [all] make-pretty markdown .md
* [ot-rfsim] make-pretty shell
* [pylibs][script] make-pretty python (excluding generates files in proto/)
* [.github] Updated Github yml workflow names; split OTNS and Nodes builds in build.yml; bring checkout before setup-go to avoid warnings. I.e. apply standard order as recommended.
* [script] moved cmake and ninja to requires packages -> needed now with Lint requiring a node-build.

+++ ot-rfsim/src/flash.c
@@ -38,11 +38,11 @@
 
 #include <openthread/config.h>
 #include <openthread/logging.h>
-#include "core/common/logging.hpp"
 #include <openthread/platform/flash.h>
+#include "core/common/logging.hpp"
 
-#include "lib/platform/exit_code.h"
 #include "common/debug.hpp"
+#include "lib/platform/exit_code.h"
 
 static int sFlashFd = -1;
 
@@ -86,7 +86,7 @@ void otPlatFlashInit(otInstance *aInstance)
 
     if (create)
     {
-        for (uint8_t index = 0; index < SWAP_NUM; index++)
+        for (uint8_t index = 0; index < (uint8_t)SWAP_NUM; index++)
         {
             otPlatFlashErase(aInstance, index);
         }
